Update and rename test.js to index.js
